+++ example/integration_test/es_test.dart
@@ -30,7 +30,7 @@ void main() async {
   EzConfig.init(
     preferences: prefs,
     defaults: testConfig,
-    fallbackLang: await EFUILang.delegate.load(english),
+    fallbackLang: await EFUILang.delegate.load(americanEnglish),
     assetPaths: <String>{},
   );