+++ example/integration_test/maximum_config_test.dart
@@ -25,7 +25,7 @@ void main() async {
   EzConfig.init(
     preferences: prefs,
     defaults: empathetechMaxConfig,
-    fallbackLang: await EFUILang.delegate.load(english),
+    fallbackLang: await EFUILang.delegate.load(americanEnglish),
     assetPaths: <String>{},
   );