+++ example/lib/main.dart
@@ -30,7 +30,7 @@ void main() async {
   EzConfig.init(
     preferences: prefs,
     defaults: isMobile() ? mobileEmpathConfig : desktopEmpathConfig,
-    fallbackLang: await EFUILang.delegate.load(english),
+    fallbackLang: await EFUILang.delegate.load(americanEnglish),
     assetPaths: <String>{},
   );